#38 refactor package
